requirements.txt
customtkinter>=5.2.2
pillow>=10.4.0
CTkColorPicker>=0.9.0
numpy>=1.26.0
streamlit>=1.35.0
